+++ src/rgw/rgw_lc.h
@@ -16,6 +16,7 @@
 #include "common/Thread.h"
 #include "rgw_common.h"
 #include "rgw_rados.h"
+#include "rgw_multi.h"
 #include "cls/rgw/cls_rgw_types.h"
 
 using namespace std;
@@ -69,6 +70,7 @@ protected:
   string status;
   LCExpiration expiration;
   LCExpiration noncur_expiration;
+  LCExpiration mp_expiration;
 
 public:
 
@@ -96,6 +98,10 @@ public:
     return noncur_expiration;
   }
 
+  LCExpiration& get_mp_expiration() {
+    return mp_expiration;
+  }
+
   void set_id(string*_id) {
     id = *_id;
   }
@@ -116,25 +122,33 @@ public:
     noncur_expiration = *_noncur_expiration;
   }
 
+  void set_mp_expiration(LCExpiration* _mp_expiration) {
+    mp_expiration = *_mp_expiration;
+  }
+
   bool validate();
   
   void encode(bufferlist& bl) const {
-     ENCODE_START(2, 1, bl);
+     ENCODE_START(3, 1, bl);
      ::encode(id, bl);
      ::encode(prefix, bl);
      ::encode(status, bl);
      ::encode(expiration, bl);
      ::encode(noncur_expiration, bl);
+     ::encode(mp_expiration, bl);
      ENCODE_FINISH(bl);
    }
    void decode(bufferlist::iterator& bl) {
-     DECODE_START_LEGACY_COMPAT_LEN(2, 1, 1, bl);
+     DECODE_START_LEGACY_COMPAT_LEN(3, 1, 1, bl);
      ::decode(id, bl);
      ::decode(prefix, bl);
      ::decode(status, bl);
      ::decode(expiration, bl);
      if (struct_v >=2) {
        ::decode(noncur_expiration, bl);
+     }
+     if (struct_v >= 3) {
+       ::decode(mp_expiration, bl);
      }
      DECODE_FINISH(bl);
    }
@@ -147,8 +161,9 @@ struct lc_op
   bool status;
   int expiration;
   int noncur_expiration;
+  int mp_expiration;
 
-  lc_op() : status(false), expiration(0), noncur_expiration(0) {}
+  lc_op() : status(false), expiration(0), noncur_expiration(0), mp_expiration(0) {}
   
 };
 
@@ -255,6 +270,7 @@ class RGWLC {
   private:
   int remove_expired_obj(RGWBucketInfo& bucket_info, rgw_obj_key obj_key, bool remove_indeed = true);
   bool obj_has_expired(double timediff, int days);
+  int handle_multipart_expiration(RGWRados::Bucket *target, const map<string, lc_op>& prefix_map);
 };
